finished ch5 exercises

+++ DogGo/Controllers/DogController.cs
@@ -1,7 +1,9 @@
 ﻿using DogGo.Models;
 using DogGo.Repositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace DogGo.Controllers
 {
@@ -13,9 +15,12 @@ namespace DogGo.Controllers
             _dogRepo = dogRepository;
         }
         // GET: DogController
+        [Authorize]
         public ActionResult Index()
         {
-            List<Dog> dogs = _dogRepo.GetAllDogs();
+            int ownerId = GetCurrentUserId();
+
+            List<Dog> dogs = _dogRepo.GetDogsByOwnerId(ownerId);
 
             return View(dogs);
         }
@@ -29,6 +34,7 @@ namespace DogGo.Controllers
         }
 
         // GET: DogController/Create
+        [Authorize]
         public ActionResult Create()
         {
             return View();
@@ -41,6 +47,9 @@ namespace DogGo.Controllers
         {
             try
             {
+                // update the dogs OwnerId to the current user's Id
+                dog.OwnerId = GetCurrentUserId();
+
                 _dogRepo.AddDog(dog);
 
                 return RedirectToAction("Index");
@@ -50,13 +59,14 @@ namespace DogGo.Controllers
                 return View(dog);
             }
         }
-
         // GET: DogController/Edit/5
+        [Authorize]
         public ActionResult Edit(int id)
         {
             Dog dog = _dogRepo.GetDogById(id);
+            int LoggedInId = GetCurrentUserId();
 
-            if (dog == null)
+            if (dog == null || dog.OwnerId != LoggedInId)
             {
                 return NotFound();
             }
@@ -69,22 +79,39 @@ namespace DogGo.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit(int id, Dog dog)
         {
-            try
+            int LoggedInId = GetCurrentUserId();
+            if (dog.OwnerId == LoggedInId)
             {
-                _dogRepo.UpdateDog(dog);
 
-                return RedirectToAction("Index");
+                try
+                {
+                    _dogRepo.UpdateDog(dog);
+
+                    return RedirectToAction("Index");
+                }
+                catch (Exception ex)
+                {
+                    return View(dog);
+                }
             }
-            catch (Exception ex)
+            else
             {
                 return View(dog);
             }
         }
 
         // GET: DogController/Delete/5
+        [Authorize]
         public ActionResult Delete(int id)
         {
+            int LoggedInId = GetCurrentUserId();
+
             Dog dog = _dogRepo.GetDogById(id);
+
+            if (dog == null || dog.OwnerId != LoggedInId)
+            {
+                return NotFound();
+            }
 
             return View(dog);
         }
@@ -94,16 +121,31 @@ namespace DogGo.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Delete(int id, Dog dog)
         {
-            try
-            {
-                _dogRepo.DeleteDog(id);
+            // get current user's Id
+            int LoggedInId = GetCurrentUserId();
 
-                return RedirectToAction("Index");
+            if (dog.OwnerId != LoggedInId)
+            {
+                try
+                {
+                    _dogRepo.DeleteDog(id);
+
+                    return RedirectToAction("Index");
+                }
+                catch (Exception ex)
+                {
+                    return View(dog);
+                }
             }
-            catch (Exception ex)
+            else
             {
                 return View(dog);
             }
+        }
+        private int GetCurrentUserId()
+        {
+            string id = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            return int.Parse(id);
         }
     }
 }
